+++ ngsindex/cli.py
@@ -1,13 +1,13 @@
 import argparse
 import ngsindex
 import json
-from xphyle import open_
+from xphyle import STDOUT, open_
 
 
 def summarize():
     parser = argparse.ArgumentParser()
     parser.add_argument(
-        "-o", "--output", default="-",
+        "-o", "--output", default=STDOUT,
         help="Output file (defaults to stdout)."
     )
     parser.add_argument("index", help="Index file.")
@@ -15,5 +15,5 @@ def summarize():
 
     index = ngsindex.parse_index(args.index)
     summary = index.summarize()
-    with open_(args.output) as out:
+    with open_(args.output, "w") as out:
         json.dump(summary, out, indent=4)